Add files via upload
SlicerHelpers contains the functions discussed in the workflow document. The method "generate_backwards_slice" takes in a python file (execution log) and a line number, as well as optional parameters to name the generated script file and the backwards slice file, and will print out the backwards slice.

The testFile.py is an example test file.